몬스터 관련 기능 수정
1. 시간에 비례해 몬스터 강해짐
- 스텟을 정하는건 스폰 시기에 정해짐
- 10분에 2배까지 강해지게 설정

2. 일반 몬스터 사망 애니메이션 설정

3. 몬스터 별로 가중치 재조정

4. 코스트 기능 삭제

5. 각 몬스터 애니메이션 정비

+++ Assets/02. Scrpits/Enemy/EnemyS/Boss/Stage1Boss/Stage1Boss.cs
@@ -39,7 +39,7 @@ public class Stage1Boss : BossBase
         StartCoroutine(LaserSpin());
     }
 
-    public override void TakeDamage(float amount)
+    public override void TakeDamage(int amount)
     {
         Stat.curHp -= amount;
         if (Stat.curHp < (Stat.totalHp * 0.8f) && !isFirstDrop)
